Extract resource manager related controller support methods to new concern
